completed auth

+++ internal/model/user.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct {
-	Id       int    `json:"id" db:"id"`
+	Id       int64  `json:"id" db:"id"`
 	Name     string `json:"name" validate:"required"`
 	Username string `json:"username" validate:"required"`
 	Password string `json:"password" validate:"required"`
